+++ covalent_ui/api/v1/utils/file_name.py
@@ -25,10 +25,3 @@ class FileName(Enum):
     """File names"""
 
     RESULTS = "result"
-
-
-class FileAttributes(Enum):
-    """File Attributes"""
-
-    DOC = "doc"
-    NAME = "name"