+++ covalent_ui/webapp/src/components/common/NodeDrawer.js
@@ -133,7 +133,13 @@ const NodeDrawer = ({ node, dispatchId }) => {
           boxSizing: 'border-box',
           border: 'none',
           p: 3,
+          marginRight: '10px',
+          marginTop:'22px',
+          height: '904px',
           bgcolor: alpha(theme.palette.background.default),
+          boxShadow:'0px 16px 50px rgba(0, 0, 0, 0.9)',
+          backdropFilter:'blur(8px)',
+          borderRadius: '16px',
         },
       })}
       anchor="right"
@@ -194,14 +200,14 @@ const NodeDrawer = ({ node, dispatchId }) => {
           {electronErrorData && <ErrorCard error={electronErrorData.data} />}
 
           {/* Description */}
-          {electronDetail.doc &&
+          {electronDetail.description &&
             (electronDetailIsFetching ? (
               <Skeleton data-testid="node__desc_skl" />
             ) : (
               <>
                 <Heading>Description</Heading>
                 <Typography fontSize="body2.fontSize" color="text.tertiary">
-                  {electronDetail.doc}
+                  {electronDetail.description}
                 </Typography>
               </>
             ))}